# doc/api.yaml
openapi: 3.0.3
info:
  title: WASAphoto
  description: |-
    Keep in touch with your friends by sharing photos of special moments, thanks to WASAPhoto! You can upload your photos directly from your PC, and they will be visible to everyone following you.

    ### Functional design specifications
    Each user will be presented with a stream of photos (images) in reverse chronological order, with information about when each photo was uploaded (date and time) and how many likes and comments it has. The stream is composed by photos from “following” (other users that the user follows).

    Users can place (and later remove) a "like" to photos from other users. Also, users can add comments to any image (even those uploaded by themself). Only authors can remove their comments.

    Users can ban other users. If user Alice bans user Eve, Eve won't be able to see any information about Alice. Alice can decIde to remove the ban at any moment.

    Users will have their profiles. The personal profile page for the user shows: the user's photos (in reverse chronological order), how many photos have been uploaded, and the user's followers and following.

    Users can change their usernames, upload photos, remove photos, and follow/unfollow other users. Removal of an image will also remove likes and comments.

    A user can search other user profiles via username.

    A user can log in just by specifying the username. See the "Simplified login" section for details.
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT    # optional, arbitrary value for documentation purposes
  responses:
    NoContent: #204
      description: The request was successful, but the client doesn't need to navigate away from its current page.
    BadRequest: #400
      description: The request was not compliant with the documentation (eg. missing fields, etc).
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Unauthorized: #401
      description: The request was not completed due to lack of authentication credentials for the requested resource.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    NotFound: #404
      description: The specified resource was not found.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    InternalServerError: #500
      description: The server encountered an internal error. Further info in server logs.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
  schemas:
    Error:
      title: Error
      description: |-
        This object represents a single error to be returned for certain requests
      type: object
      properties:
        code:
          type: string
          description: An HTTP error code
          example: "404"
        message:
          type: string
          description: A brief message describing the error code
          example: The specified resource was not found
      required: [ "code", "message" ]
...
    Profile:
      title: Profile
      description: This is the user's profile
      type: object
      properties:
        posts:
          title: photos
          description: This contains all of the user's posts
          type: array
          items: {$ref: '#/components/schemas/Post'}
          minItems: 0
          maxItems: 99999
        numberOfPhotos:
          title: number of photos
          description: this is the number of photos uploaded by a user
          type: integer
          example: 1
        followers:
          title: followers
          description: This contains all of the user's followers
          type: array
          items: {$ref: '#/components/schemas/User/properties/username'}
          minItems: 0
          maxItems: 99999
        followedUsers:
          title: following
          description: This contains all of the user's followed users
          type: array
          items: {$ref: '#/components/schemas/User/properties/username'}
          minItems: 0
          maxItems: 99999
    Comment:
      title: Comment
      description: This is a User's comment on a certain photo
      type: object
      properties:
        commentOwner: {$ref: '#/components/schemas/User/properties/username'}
        commentText:
          title: comment Text
          description: This is the content of the comment
          type: string
          pattern: '^.*$'
          minLength: 1
          maxLength: 2200 #Like instagram
          example: "Good photo!"
        creationTime:
          title: creation Time
          description: this is the creation time and date
          type: string
          pattern: '^.*$'
          minLength: 1
          maxLength: 99999
          format: date-time
          example: 2017-07-21T17:32:28Z
        commentId:
          title: Comment Id
          description: This is the post's unique Id
          type: string
          pattern: '^.*$'
          minLength: 1
          maxLength: 99999
          example: "Comment123"
    Like:
      title: like
      description: like object
      properties:
        likeOwner: { $ref: '#/components/schemas/User/properties/username' }
        likeId:
          title: Like Id
          description: This is the like's unique Id
          type: string
          pattern: '^.*$'
          minLength: 1
          maxLength: 99999
          example: "Like123"
        creationTime:
          title: creation Time
          description: this is the creation time and date
          type: string
          pattern: '^.*$'
          minLength: 1
          maxLength: 99999
          format: date-time
          example: 2017-07-21T17:32:28Z
    Post:
      title: Post
      description: This is a post generated by a user using a certain image
      type: object
      properties:
        postOwner: {$ref: '#/components/schemas/User/properties/username'}
        image:
          title: Image URL
          type: string
          pattern: '^.*$'
          minLength: 1
          maxLength: 99999
          description: This is the image's unique URL that points to the file
          example: "https://example.com/example/images/image.png"
        comments:
          description: array of the Ids of the comments on a certain post
          type: array
          items: {$ref: '#/components/schemas/Comment'}
          minItems: 0
          maxItems: 99999
        nComments:
          description: number of comments on a post
          type: integer
          example: 1
          minimum: 0
          maximum: 99999
        likesNumber:
          description: number of likes in a post
          type: integer
          example: 1
          minimum: 0
          maximum: 99999
        likes:
          description: array of users who put like on a post
          type: array
          items: {$ref: "#/components/schemas/Like"}
          minItems: 0
          maxItems: 99999
        creationTime:
          title: creation Time
          description: this is the creation time and date
          type: string
          pattern: '^.*$'
          minLength: 1
          maxLength: 99999
          format: date-time
          example: 2017-07-21T17:32:28Z
        postId:
          title: Post Id
          description: This is the post's unique Id
          type: string
          pattern: '^.*$'
          minLength: 1
          maxLength: 99999
          example: "Post123"
security:
  - bearerAuth: [ ]
...
paths:
  /session:
    post:
      tags: ["login"]
      summary: Logs in the user
      description: |-
        If the user does not exist, it will be created,
        and an Identifier is returned.
        If the user exists, the user Identifier is returned.
      operationId: doLogin
      requestBody:
        description: User details
        content:
          application/json:
            schema:
              description: the user's username
              type: object
              properties:
                name:
                  description: The string of the user's username
                  type: string
                  example: "Maria"
                  pattern: '^.*$'
                  minLength: 3
                  maxLength: 16
        required: true
      responses:
        "201":
          description: User log-in action successful
          content:
            application/json:
              schema:
                type: object
                description: the unique Id of the session
                properties:
                  Identifier:
                    description: The identifier
                    type: string
                    pattern: '^.*$'
                    minLength: 1
                    maxLength: 99999
                    example: "abcdef012345"
      security: []
...
  /users/profiles/{username}:
    parameters:
      - name: username
        in: path
        description: The username of the user whose profile is to be shown
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/User/properties/username' }
    get:
      tags: [ "theirProfile" ]
      operationId: getUserProfile
      summary: Retrieves the user's profile
      description: |-
        Retrieves the user's profile from the system and
        returns it to the user.
      responses:
        "200":
          description: Successfully retrieved the user's profile
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Profile' }
        "400": { $ref: '#/components/responses/BadRequest' }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "404": { $ref: '#/components/responses/NotFound' }
        "500": { $ref: '#/components/responses/InternalServerError' }
...
  /users/{username}/banned:
    parameters:
      - name: username
        in: path
        description: The username of the user who wants to ban
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/User/properties/username' }
      - name: bannedUser
        in: query
        description: The username of the user to be banned
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/User/properties/username' }
    put:
      tags: [ "theirProfile" ]
      operationId: banUser
      summary: ban a user
      description: "Given the username of a user, it allows you to ban a certain user"
      responses:
        "201":
          description: Successfully banned the user
          content:
            application/json:
              schema: { $ref: "#/components/schemas/User/properties/username" }
        "400": { $ref: '#/components/responses/BadRequest' }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "404": { $ref: '#/components/responses/NotFound' }
        "500": { $ref: '#/components/responses/InternalServerError' }
...
  /users/{username}/posts/{postId}:
    parameters:
      - name: postId
        in: path
        description: the Id of the post being liked or unliked or the post to be deleted
        required: true
        schema: { $ref: '#/components/schemas/Post/properties/postId' }
      - name: username
        in: path
        description: the username of the user whose post is being shown
        required: true
        schema: { $ref: '#/components/schemas/User/properties/username' }
    delete:
      tags: [ "myPost" ]
      description: deletes a post
      operationId: deletePhoto
      summary: delete a post
      responses:
        "204": { $ref: '#/components/responses/NoContent' }
        "400": { $ref: '#/components/responses/BadRequest' }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "404": { $ref: '#/components/responses/NotFound' }
        "500": { $ref: '#/components/responses/InternalServerError' }
  /users/{username}/posts/{postId}/likes:
    parameters:
        - name: postId
          in: path
          description: the Id of the post being liked or unliked or the post to be deleted
          required: true
          schema: {$ref: '#/components/schemas/Post/properties/postId'}
        - name: username
          in: path
          description: the username of the user whose post is being shown
          required: true
          schema: {$ref: '#/components/schemas/User/properties/username'}
    post:
      tags: ["Post"]
      operationId: likePhoto
      summary: like a post
      description: Given the postId of a post, allows you to like it
      requestBody:
        description: the username of the user who puts like to the post
        required: true
        content:
          application/json:
            schema:
              type: object
              description: the username of the user who liked
              properties:
                targetUser: {$ref: "#/components/schemas/Like"}
      responses:
        "201":
          description: Successfully liked the post
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Like" }
        "204": { $ref: '#/components/responses/NoContent' }
        "400": { $ref: '#/components/responses/BadRequest' }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "404": { $ref: '#/components/responses/NotFound' }
        "500": { $ref: '#/components/responses/InternalServerError' }
...
  /users/{username}/posts/{postId}/comments:
    post:
      tags: ["Post"]
      operationId: commentPhoto
      description: given a postId and the username of the user who created it, it allows you to write a comment on the post
      summary: write a comment ona  post
      parameters:
        - name: postId
          in: path
          description: the Id of the post being commented on
          required: true
          schema: {$ref: '#/components/schemas/Post/properties/postId'}
        - name: username
          in: path
          description: the username of the user whose post is being commented on
          required: true
          schema: {$ref: '#/components/schemas/User/properties/username'}
      requestBody:
        description: The text of the comment
        required: true
        content:
          application/json:
            schema:
              type: object
              description: The comment text schema
              properties:
                content: {$ref: '#/components/schemas/Comment/properties/commentText'}
      responses:
        "201":
          description: Successfully commented the post
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Comment/properties/commentId" }
        "204": { $ref: '#/components/responses/NoContent' }
        "400": { $ref: '#/components/responses/BadRequest' }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "404": { $ref: '#/components/responses/NotFound' }
        "500": { $ref: '#/components/responses/InternalServerError' }
...
  /users/{username}:
    parameters:
        - name: username
          in: path
          description: the current username of the user who wants to change username or look for another user
          required: true
          schema: { $ref: '#/components/schemas/User/properties/username' }
    put:
      description: it allows you to set the username
      tags: ["myProfile"]
      operationId: setMyUserName
      summary: set the username
      requestBody:
        description: Username to set
        required: true
        content:
          application/json:
            schema:
              type: object
              description: the username schema
              properties:
                  content: {$ref: '#/components/schemas/User/properties/username'}
      responses:
        "200":
          description: Successfully set the username
          content:
            application/json:
              schema: { $ref: "#/components/schemas/User/properties/username" }
        "400": { $ref: '#/components/responses/BadRequest' }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "404": { $ref: '#/components/responses/NotFound' }
        "409":
          description: Username already exists
          content:
            application/json:
              schema: { $ref: "#/components/schemas/User/properties/username" }
        "500": { $ref: '#/components/responses/InternalServerError' }
...